finished all exercise

+++ rustlings/exercises/conversions/using_as.rs
@@ -1,15 +1,6 @@
-// Type casting in Rust is done via the usage of the `as` operator.
-// Please note that the `as` operator is not only used when type casting.
-// It also helps with renaming imports.
-//
-// The goal is to make sure that the division does not fail to compile
-// and returns the proper type.
-
-// I AM NOT DONE
-
 fn average(values: &[f64]) -> f64 {
     let total = values.iter().sum::<f64>();
-    total / values.len()
+    total / values.len() as f64  // Явне приведення типу
 }
 
 fn main() {
